chore: factor out request, response, and other type definitions
