+++ packages/runtime-sdk/src/workers/_workers.py
@@ -3,723 +3,33 @@
 # programming language.
 import functools
 import inspect
-import json
 from asyncio import create_task, gather
-from collections.abc import (
-    Awaitable,
-    Generator,
-    Iterable,
-    MutableMapping,
-)
-from contextlib import ExitStack, contextmanager
-from enum import StrEnum
-from http import HTTPMethod, HTTPStatus
-from typing import TYPE_CHECKING, Any, Never, Protocol, TypedDict, Unpack
+from typing import TYPE_CHECKING, Any
 
 import _cloudflare_compat_flags
 
 # Get globals modules and import function from the entrypoint-helper
 import _pyodide_entrypoint_helper
 import js
-import pyodide.http
 from js import Object
-from pyodide import __version__ as pyodide_version
 from pyodide.ffi import (
-    JsBuffer,
-    JsException,
     JsProxy,
     create_once_callable,
-    create_proxy,
     to_js,
 )
-from pyodide.http import pyfetch
 
-from .rpc import (
-    python_from_rpc,
-    python_to_rpc,
-)
+from .fetch import fetch
+from .request import Request
+from .rpc import python_from_rpc, python_to_rpc
 from .utils import (
     _JS_PASSTHROUGH_TYPES,
-    RESPONSE_ACCEPTED_TYPES,
     _from_js_error,
     _get_js_constructor_name,
-    _is_iterable,
     _is_js_instance,
-    _js_headers_to_http_message,
-    _jsnull_to_none,
-    _manage_pyproxies,
-    _supports_buffer_protocol,
-    _to_js_headers,
-    _to_python_exception,
 )
 
 if TYPE_CHECKING:
     from js import DurableObjectState, Env, ExecutionContext
-
-
-class Context(Protocol):
-    def waitUntil(self, other: Awaitable[Any]) -> None: ...
-
-
-type JSBody = (
-    "js.Blob | JsBuffer | js.FormData | js.ReadableStream | js.URLSearchParams"
-)
-type Body = "str | FormData | JSBody"
-type Headers = "dict[str, str] | list[tuple[str, str]] | js.Headers"
-
-
-# https://developers.cloudflare.com/workers/runtime-apis/request/#the-cf-property-requestinitcfproperties
-class RequestInitCfProperties(TypedDict, total=False):
-    apps: bool | None
-    cacheEverything: bool | None
-    cacheKey: str | None
-    cacheTags: list[str] | None
-    cacheTtl: int
-    cacheTtlByStatus: dict[str, int]
-    image: (
-        Any | None
-    )  # TODO: https://developers.cloudflare.com/images/transform-images/transform-via-workers/
-    mirage: bool | None
-    polish: str | None
-    resolveOverride: str | None
-    scrapeShield: bool | None
-    webp: bool | None
-
-
-# This matches the Request options:
-# https://developers.cloudflare.com/workers/runtime-apis/request/#options
-class FetchKwargs(TypedDict, total=False):
-    headers: "Headers | None"
-    body: "Body | None"
-    method: HTTPMethod | None
-    redirect: str | None
-    cf: RequestInitCfProperties | None
-    fetcher: type[pyfetch] | None
-
-
-class FetchResponse(pyodide.http.FetchResponse):
-    # TODO: Consider upstreaming the `body` attribute
-    # TODO: Behind a compat flag make this return a native stream (StreamReader?), or perhaps
-    #       behind a different name, maybe `stream`?
-    @property
-    def body(self) -> "js.ReadableStream":
-        """
-        Returns the body as a JavaScript ReadableStream from the JavaScript Response instance.
-        """
-        return _jsnull_to_none(self.js_response.body)
-
-    @property
-    def js_object(self) -> "js.Response":
-        return self.js_response
-
-    @property
-    def headers(self):
-        return _js_headers_to_http_message(self.js_object.headers)
-
-    """
-    Instance methods defined below.
-
-    Some methods are implemented by `FetchResponse`, these include `buffer`
-    (replacing JavaScript's `arrayBuffer`), `bytes`, `json`, and `text`.
-
-    There are also some additional methods implemented by `FetchResponse`.
-    See https://pyodide.org/en/stable/usage/api/python-api/http.html#pyodide.http.FetchResponse
-    for details.
-    """
-
-    async def formData(self) -> "FormData":  # TODO: Remove after certain compat date.
-        return await self.form_data()
-
-    async def form_data(self) -> "FormData":
-        self._raise_if_failed()
-        try:
-            return FormData(await self.js_response.formData())
-        except JsException as exc:
-            raise _to_python_exception(exc) from exc
-
-    def replace_body(self, body: Body) -> "Response":
-        """
-        Returns a new Response object with the same options (status, headers, etc) as
-        the original but with an updated body.
-        """
-        b = body.js_object if isinstance(body, FormData) else body
-        js_resp = js.Response.new(b, self.js_response)
-        return Response(js_resp)
-
-    async def blob(self) -> "Blob":
-        self._raise_if_failed()
-        return Blob(await self.js_object.blob())
-
-    """
-    Static methods defined below. The `error` static method is not implemented as
-    it is not useful for the Workers use case.
-    """
-
-    @staticmethod
-    def redirect(url: str, status: HTTPStatus | int = HTTPStatus.FOUND):
-        code = status.value if isinstance(status, HTTPStatus) else status
-        try:
-            return js.Response.redirect(url, code)
-        except JsException as exc:
-            raise _to_python_exception(exc) from exc
-
-    @staticmethod
-    def from_json(
-        data: str | dict[str, Any] | list[Any] | JsProxy,
-        status: HTTPStatus | int = HTTPStatus.OK,
-        status_text="",
-        headers: Headers = None,
-    ) -> "Response":
-        options = Response._create_options(status, status_text, headers)
-        js_resp = None
-        try:
-            if isinstance(data, JsProxy):
-                js_resp = js.Response.json(data, **options)
-            else:
-                if "headers" not in options:
-                    options["headers"] = _to_js_headers(
-                        {"content-type": "application/json"}
-                    )
-                elif not options["headers"].has("content-type"):
-                    options["headers"].set("content-type", "application/json")
-                js_resp = js.Response.new(json.dumps(data), **options)
-        except JsException as exc:
-            raise _to_python_exception(exc) from exc
-
-        return Response(js_resp)
-
-    def json(self, *args: Never, **kwargs: Never):
-        if isinstance(self, Response):
-            return super().json()
-        # For compatibility, allow static use of Response.json() to mean Response.from_json().
-        data = self
-        return Response.from_json(data, *args, **kwargs)
-
-
-if pyodide_version == "0.26.0a2":
-
-    async def _pyfetch_patched(
-        request: "str | js.Request", **kwargs: Any
-    ) -> "Response":
-        # This is copied from https://github.com/pyodide/pyodide/blob/d3f99e1d/src/py/pyodide/http.py
-        custom_fetch = kwargs["fetcher"] if "fetcher" in kwargs else js.fetch
-        kwargs["fetcher"] = None
-        try:
-            return Response(
-                await custom_fetch(
-                    request, to_js(kwargs, dict_converter=Object.fromEntries)
-                ),
-            )
-        except JsException as e:
-            raise OSError(e.message) from None
-else:
-    _pyfetch_patched = pyfetch
-
-
-async def fetch(
-    resource: "str | Request | js.Request",
-    **other_options: Unpack[FetchKwargs],
-) -> "Response":
-    if isinstance(resource, Request):
-        resource = resource.js_object
-    if "method" in other_options and isinstance(other_options["method"], HTTPMethod):
-        other_options["method"] = other_options["method"].value
-
-    resp = await _pyfetch_patched(resource, **other_options)
-    return Response(resp.js_response)
-
-
-class Response(FetchResponse):
-    """
-    This class represents the response to an HTTP request, with a similar API to that of the web
-    `Response` API: https://developer.mozilla.org/en-US/docs/Web/API/Response.
-    """
-
-    def __init__(
-        self,
-        body: Body = None,
-        status: HTTPStatus | int | None = None,
-        status_text="",
-        headers: Headers = None,
-        web_socket: "js.WebSocket | None" = None,
-    ):
-        """
-        Represents the response to a request.
-
-        Based on the JS API of the same name:
-        https://developer.mozilla.org/en-US/docs/Web/API/Response/Response.
-        """
-        # Verify passed in types.
-        js_type = _get_js_constructor_name(body)
-        if js_type:
-            if js_type not in RESPONSE_ACCEPTED_TYPES:
-                raise TypeError(f"Unsupported type in Response: {js_type}")
-        elif not isinstance(body, str | FormData | bytes) and body is not None:
-            raise TypeError(f"Unsupported type in Response: {type(body).__name__}")
-
-        # Handle constructing a Response from a JS Response.
-        if _is_js_instance(body, "Response"):
-            if status is not None or len(status_text) > 0 or headers is not None:
-                raise ValueError(
-                    "Expected no options when constructing Response from a js.Response"
-                )
-            super().__init__(body.url, body)
-            return
-
-        options = self._create_options(status, status_text, headers, web_socket)
-
-        # To avoid unnecessary copies we use this context manager.
-        with _get_js_body(body) as js_body:
-            # Initialize via the FetchResponse super-class which gives us access to
-            # methods that we would ordinarily have to redeclare.
-            js_resp = js.Response.new(js_body, **options)
-        super().__init__(js_resp.url, js_resp)
-
-    def __repr__(self):
-        body = [f"status={self.status}"]
-        if self.js_object.statusText:
-            body.append(f"status_text={self.status_text!r}")
-        if "content-type" in self.headers:
-            body.append(f"content_type={self.headers['content-type']!r}")
-        if self.js_object.url:
-            body.append(f"url={self.js_object.url!r}")
-        if self.js_object.type != "default":
-            body.append(f"type={self.js_object.type!r}")
-        return f"Response({', '.join(body)})"
-
-    @staticmethod
-    def _create_options(
-        status: HTTPStatus | int | None = HTTPStatus.OK,
-        status_text="",
-        headers: Headers = None,
-        web_socket: "js.WebSocket | None" = None,
-    ):
-        options = {}
-        if status:
-            options["status"] = (
-                status.value if isinstance(status, HTTPStatus) else status
-            )
-        if status_text:
-            options["statusText"] = status_text
-        if headers:
-            options["headers"] = _to_js_headers(headers)
-        if web_socket:
-            options["webSocket"] = web_socket
-        return options
-
-
-FormDataValue = "str | js.Blob | Blob"
-
-
-def _py_value_to_js(item: FormDataValue) -> "str | js.Blob":
-    if isinstance(item, Blob):
-        return item.js_object
-    else:
-        return item
-
-
-def _js_value_to_py(item: FormDataValue) -> "str | Blob | File":
-    if hasattr(item, "constructor") and (item.constructor.name in ("Blob", "File")):
-        if item.constructor.name == "File":
-            return File(item, item.name)
-        else:
-            return Blob(item)
-    else:
-        return item
-
-
-class FormData(MutableMapping[str, FormDataValue]):
-    """
-    This class represents a set of key/value pairs for forms.
-
-    The API of this class follows that of https://pypi.org/project/multidict/ and
-    https://developer.mozilla.org/en-US/docs/Web/API/FormData.
-    """
-
-    def __init__(
-        self, form_data: "js.FormData | None | dict[str, FormDataValue]" = None
-    ):
-        if not form_data:
-            self._js_form_data = js.FormData.new()
-            return
-
-        if isinstance(form_data, dict):
-            self._js_form_data = js.FormData.new()
-            for k, v in form_data.items():
-                self._js_form_data.append(k, _py_value_to_js(v))
-            return
-
-        if _is_js_instance(form_data, "FormData"):
-            self._js_form_data = form_data
-            return
-
-        raise TypeError("Expected form_data to be a dict or an instance of FormData")
-
-    def __getitem__(self, key: str) -> FormDataValue:
-        return _js_value_to_py(self._js_form_data.get(key))
-
-    def __setitem__(self, key: str, value: FormDataValue):
-        if isinstance(value, list):
-            raise TypeError("Expected single item in arguments to FormData.__setitem__")
-        self._js_form_data.set(key, _py_value_to_js(value))
-
-    def append(self, key: str, value: FormDataValue, filename: str | None = None):
-        self._js_form_data.append(key, _py_value_to_js(value), filename)
-
-    def delete(self, key: str):
-        self._js_form_data.delete(key)
-
-    def __contains__(self, key: str) -> bool:
-        return self._js_form_data.has(key)
-
-    def values(self) -> Generator[FormDataValue, None, None]:
-        for val in self._js_form_data.values():
-            yield _js_value_to_py(val)
-
-    def keys(self) -> Generator[str, None, None]:
-        yield from self._js_form_data.keys()
-
-    def __iter__(self):
-        yield from self.keys()
-
-    def items(self) -> Generator[tuple[str, FormDataValue], None, None]:
-        for k, v in self._js_form_data.entries():
-            yield (k, _js_value_to_py(v))
-
-    def __delitem__(self, key: str):
-        self.delete(key)
-
-    def __len__(self):
-        return len(self.keys())
-
-    def get_all(self, key: str) -> list[FormDataValue]:
-        return [_js_value_to_py(x) for x in self._js_form_data.getAll(key)]
-
-    @property
-    def js_object(self) -> "js.FormData":
-        return self._js_form_data
-
-
-@contextmanager
-def _get_js_body(body):
-    if isinstance(body, bytes):
-        proxy_bytes = create_proxy(body)
-        proxy_buffer = proxy_bytes.getBuffer()
-        try:
-            yield proxy_buffer.data
-            return
-        finally:
-            proxy_buffer.release()
-            proxy_bytes.destroy()
-    if isinstance(body, FormData):
-        yield body.js_object
-        return
-    yield body
-
-
-@contextmanager
-def _make_blob_entry(e):
-    if isinstance(e, str):
-        yield e
-        return
-    if isinstance(e, Blob):
-        yield e._js_blob
-        return
-    if hasattr(e, "constructor") and (e.constructor.name in ("Blob", "File")):
-        yield e
-        return
-    if _supports_buffer_protocol(e):
-        px = create_proxy(e)
-        buf = px.getBuffer()
-        try:
-            yield buf.data
-            return
-        finally:
-            buf.release()
-            px.destroy()
-    raise TypeError(f"Don't know how to handle {type(e)} for Blob()")
-
-
-BlobValue = (
-    "str | bytes | js.ArrayBuffer | js.TypedArray | js.DataView | js.Blob | Blob | File"
-)
-
-
-class BlobEnding(StrEnum):
-    TRANSPARENT = "transparent"
-    NATIVE = "native"
-
-
-class Blob:
-    def __init__(
-        self,
-        blob_parts: "Iterable[BlobValue] | BlobValue",
-        content_type: str | None = None,
-        endings: BlobEnding | str | None = None,
-    ):
-        if endings:
-            endings = str(endings)
-
-        is_single_item = not _is_iterable(blob_parts)
-        if is_single_item:
-            # Inherit the content_type if we have a single item. If a File is passed
-            # in then its metadata is lost.
-            if not content_type and isinstance(blob_parts, Blob):
-                content_type = blob_parts.content_type
-            if hasattr(blob_parts, "constructor") and (
-                blob_parts.constructor.name in ("Blob", "File")
-            ):
-                if not content_type:
-                    content_type = blob_parts.type
-
-            # Otherwise create a new Blob below.
-            blob_parts = [blob_parts]
-
-        with ExitStack() as stack:
-            args = [stack.enter_context(_make_blob_entry(e)) for e in blob_parts]
-            with _manage_pyproxies() as pyproxies:
-                self._js_blob = js.Blob.new(
-                    to_js(args, pyproxies=pyproxies),
-                    type=content_type,
-                    endings=endings,
-                )
-
-    @property
-    def size(self) -> int:
-        return self._js_blob.size
-
-    @property
-    def content_type(self) -> str:
-        return self._js_blob.type
-
-    @property
-    def js_object(self) -> "js.Blob":
-        return self._js_blob
-
-    async def text(self) -> str:
-        return await self.js_object.text()
-
-    async def bytes(self) -> bytes:
-        return (await self.js_object.arrayBuffer()).to_bytes()
-
-    def slice(
-        self,
-        start: int | None = None,
-        end: int | None = None,
-        content_type: str | None = None,
-    ):
-        js_sliced_blob = self.js_object.slice(start, end, content_type)
-        return Blob([js_sliced_blob])
-
-
-class File(Blob):
-    def __init__(
-        self,
-        blob_parts: "Iterable[BlobValue] | BlobValue",
-        filename: str,
-        content_type: str | None = None,
-        endings: BlobEnding | str | None = None,
-        last_modified: int | None = None,
-    ):
-        if endings:
-            endings = str(endings)
-
-        is_single_item = not _is_iterable(blob_parts)
-        if is_single_item:
-            # Inherit the content_type and lastModified if we have a
-            # single item.
-            if not content_type and isinstance(blob_parts, Blob):
-                content_type = blob_parts.content_type
-            if not last_modified and isinstance(blob_parts, File):
-                last_modified = blob_parts.last_modified
-            if hasattr(blob_parts, "constructor") and (
-                blob_parts.constructor.name in ("Blob", "File")
-            ):
-                if not content_type:
-                    content_type = blob_parts.type
-                if blob_parts.constructor.name == "File":
-                    if not last_modified:
-                        last_modified = blob_parts.lastModified
-
-            # Otherwise create a new File below.
-            blob_parts = [blob_parts]
-
-        with ExitStack() as stack:
-            args = [stack.enter_context(_make_blob_entry(e)) for e in blob_parts]
-            with _manage_pyproxies() as pyproxies:
-                self._js_blob = js.File.new(
-                    to_js(args, pyproxies=pyproxies),
-                    filename,
-                    type=content_type,
-                    endings=endings,
-                    lastModified=last_modified,
-                )
-
-    @property
-    def name(self) -> str:
-        return self._js_blob.name
-
-    @property
-    def last_modified(self) -> int:
-        return self._js_blob.lastModified
-
-
-class Request:
-    def __init__(
-        self, input: "Request | str | js.Request", **other_options: Unpack[FetchKwargs]
-    ):
-        if _is_js_instance(input, "Request"):
-            if len(other_options) > 0:
-                raise ValueError(
-                    "Expected no options when constructing Request from a js.Request"
-                )
-            self._js_request = input
-            return
-
-        if "method" in other_options and isinstance(
-            other_options["method"], HTTPMethod
-        ):
-            other_options["method"] = other_options["method"].value
-
-        if "headers" in other_options:
-            other_options["headers"] = _to_js_headers(other_options["headers"])
-        self._js_request = js.Request.new(
-            input._js_request if isinstance(input, Request) else input, **other_options
-        )
-
-    def __repr__(self):
-        return (
-            f"Request(method={self._js_request.method!r}, url={self._js_request.url!r})"
-        )
-
-    @property
-    def js_object(self) -> "js.Request":
-        return self._js_request
-
-    # TODO: expose `body` as a native Python stream in the future, follow how we define `Response`
-    @property
-    def body(self) -> "js.ReadableStream":
-        return self.js_object.body
-
-    @property
-    def body_used(self) -> bool:
-        return self.js_object.bodyUsed
-
-    @property
-    def cf(self) -> "JsProxy | None":
-        """
-        Cloudflare-specific properties about the incoming request
-        (IncomingRequestCfProperties).  Access fields via attribute
-        notation, for example ``request.cf.colo``.
-
-        Returns None when not present (for example, in the dashboard/playground
-        preview or for requests constructed without a ``cf`` value).
-
-        See https://developers.cloudflare.com/workers/runtime-apis/request/#incomingrequestcfproperties
-        """
-        return _jsnull_to_none(self.js_object.cf)
-
-    @property
-    def cache(self) -> str:
-        return self.js_object.cache
-
-    @property
-    def credentials(self) -> str:
-        return self.js_object.credentials
-
-    @property
-    def destination(self) -> str:
-        return self.js_object.destination
-
-    @property
-    def headers(self):
-        return _js_headers_to_http_message(self.js_object.headers)
-
-    @property
-    def integrity(self) -> str:
-        return self.js_object.integrity
-
-    @property
-    def is_history_navigation(self) -> bool:
-        return self.js_object.isHistoryNavigation
-
-    @property
-    def keepalive(self) -> bool:
-        return self.js_object.keepalive
-
-    @property
-    def method(self) -> HTTPMethod:
-        return HTTPMethod[self.js_object.method]
-
-    @property
-    def mode(self) -> str:
-        return self.js_object.mode
-
-    @property
-    def redirect(self) -> str:
-        return self.js_object.redirect
-
-    @property
-    def referrer(self) -> str:
-        return self.js_object.referrer
-
-    @property
-    def referrer_policy(self) -> str:
-        return self.js_object.referrerPolicy
-
-    @property
-    def url(self) -> str:
-        return self.js_object.url
-
-    def _raise_if_failed(self) -> None:
-        # TODO: https://github.com/pyodide/pyodide/blob/a53c17fd8/src/py/pyodide/http.py#L252
-        if self.body_used:
-            # TODO: Use BodyUsedError in newer Pyodide versions.
-            raise OSError("Body already used")
-
-    """
-    Instance methods defined below.
-
-    The naming of these methods should match Request's methods when possible.
-
-    TODO: AbortController support.
-    """
-
-    async def buffer(self) -> "js.ArrayBuffer":
-        # The naming of this method matches that of Response.
-        self._raise_if_failed()
-        return await self.js_object.arrayBuffer()
-
-    async def form_data(self) -> "FormData":
-        self._raise_if_failed()
-        try:
-            return FormData(await self.js_object.formData())
-        except JsException as exc:
-            raise _to_python_exception(exc) from exc
-
-    async def blob(self) -> Blob:
-        self._raise_if_failed()
-        return Blob(await self.js_object.blob())
-
-    async def bytes(self) -> bytes:
-        self._raise_if_failed()
-        return (await self.buffer()).to_bytes()
-
-    def clone(self) -> "Request":
-        if self.body_used:
-            # TODO: Use BodyUsedError in newer Pyodide versions.
-            raise OSError("Body already used")
-        return Request(
-            self.js_object.clone(),
-        )
-
-    async def json(self, **kwargs: Any) -> Any:
-        self._raise_if_failed()
-        return json.loads(await self.text(), **kwargs)
-
-    async def text(self) -> str:
-        self._raise_if_failed()
-        return await self.js_object.text()
 
 
 class _BindingWrapper:
